fix problem with conf file regeneration
The conf file regeneration inside the main driver loop wasn't working.  For
some reason the loop appears to have been isolated from other transactions
occuring on the db, so it wasn't noticing the new data (i.e., the data was
there, and the loop would work, but it would always see the old data!)

Moved the conf file generation to its own script, which is run independently.
Since is starts, runs, then stops it has no way to hang on to old state and
seems to work like a charm. 


git-svn-id: http://svn.donomo.net/trunk@695 3f4560be-fcb6-4516-b323-ae570ede3bb7